move SoonPom to support/compiler
