+++ src/main/java/spoon/support/compiler/SpoonPom.java
@@ -14,7 +14,7 @@
  * The fact that you are presently reading this means that you have had
  * knowledge of the CeCILL-C license and that you accept its terms.
  */
-package spoon.internal.mavenlauncher;
+package spoon.support.compiler;
 
 import org.apache.log4j.Logger;
 import org.apache.maven.model.Model;
@@ -110,20 +110,16 @@ public class SpoonPom implements SpoonResource {
 		}
 	}
 
-	public void addModule(SpoonPom module) {
+	private void addModule(SpoonPom module) {
 		modules.add(module);
 	}
 
+	/**
+	 * Get the Project Object Model
+	 * @return the Project Object Model
+	 */
 	public Model getModel() {
 		return model;
-	}
-
-	/**
-	 * Get the parent model
-	 * @return the parent model
-	 */
-	public SpoonPom getParentPom() {
-		return parent;
 	}
 
 	/**
@@ -184,6 +180,10 @@ public class SpoonPom implements SpoonResource {
 		return output;
 	}
 
+	/**
+	 * Get the list of classpath files generated by maven
+	 * @return the list of classpath files
+	 */
 	public List<File> getClasspathTmpFiles(String fileName) {
 		List<File> output = new ArrayList<>();
 		File tmp = new File(directory, fileName);
@@ -466,6 +466,19 @@ public class SpoonPom implements SpoonResource {
 			return spoonClasspathTmpFileName;
 		}
 	}
+
+	/**
+	 * Get the parent model
+	 * @return the parent model
+	 */
+	public SpoonPom getParentPom() {
+		return parent;
+	}
+
+	/**
+	 * Get the parent directory
+	 * @return the parent directory
+	 */
 	public SpoonFolder getParent() {
 		try {
 			return SpoonResourceHelper.createFolder(directory);
