+++ src/test/java/spoon/test/architecture/SpoonArchitectureEnforcerTest.java
@@ -341,8 +341,6 @@ public class SpoonArchitectureEnforcerTest {
 		officialPackages.add("spoon.support.modelobs.context");
 		officialPackages.add("spoon.support.modelobs");
 		officialPackages.add("spoon.experimental");
-		officialPackages.add("spoon.internal");
-		officialPackages.add("spoon.internal.mavenlauncher");
 		officialPackages.add("spoon.legacy");
 		officialPackages.add("spoon.metamodel");
 		officialPackages.add("spoon.pattern");
